+++ ollama/_types.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from datetime import datetime
 from typing import (
   Any,
+  List,
   Literal,
   Mapping,
   Optional,
@@ -229,13 +230,13 @@ class Tool(SubscriptableBaseModel):
     description: str
 
     class Parameters(SubscriptableBaseModel):
-      type: Union[str, list[str]]
+      type: Union[str, List[str]]
       required: Optional[Sequence[str]] = None
       properties: Optional[JsonSchemaValue] = None
 
     parameters: Parameters
 
-    return_type: Optional[str | list[str]] = None
+    return_type: Optional[Union[str, List[str]]] = None
 
   function: Function
